Adding blank element to server dropdown
This is so that if a server being used in a job gets deleted, the job wont
point to the first server in the list.  Could be dangerous if a VM in the
deleted server has the same name as one in the first server OR if someone
just makes a mistake and forgets to save a server.