variable expansion and server check in build steps
ensure the server exists in the global config.  It's possible a server was
deleted but the job config was not updated and still holds the server
data.  It's also possible the server objects in the global config will
change while keeping the same name.  This check will prevent any
unexpected, and possibly dangerous operations.